+++ server/databasemigration_test.go
@@ -329,7 +329,7 @@ func newAppConfigurationWithDefaults() config.AppConfiguration {
 		TLSMinimumVersion: "1.2",
 	}
 	conf = config.NewAppConfiguration(opts)
-	conf.ServerSettings.AclImpersonationWhitelist = whitelist
+	conf.ServerSettings.ACLImpersonationWhitelist = whitelist
 	return conf
 }